backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
python-multipart==0.0.6
pydantic==2.5.0
opencv-python==4.8.1.78
scikit-learn==1.3.2